+++ FileSystem/src/functions/operaciones.h
@@ -11,7 +11,7 @@
 int validar(char* path);
 int crear(char* path);
 int borrar(char* path);
-size_t obtenerDatos(char* path, off_t offset, size_t size);
+int obtenerDatos(char* path, off_t offset, size_t size);
 int guardarDatos(char* path, off_t offset, size_t size, void* buffer);
 
 #endif /* FUNCTIONS_OPERACIONES_H_ */
